DPTS-519 Refactoring code and tests into common
