+++ application/utilities/common.py
@@ -1,4 +1,7 @@
+from utilities.database import DB
+from utilities.s3 import S3
 from utilities.logger import log_for_audit, log_for_error  # noqa
+from utilities.message import send_success_slack_message, send_failure_slack_message
 
 
 def check_csv_format(csv_row, csv_column_count):
@@ -20,3 +23,34 @@ def valid_action(record_exists, row_data):
     if not valid_action:
         log_for_error("Invalid action {} for the record with ID {}".format(row_data["action"], row_data["id"]))
     return valid_action
+
+
+def cleanup(db_connection, bucket, filename, event, start):
+    # Close DB connection
+    log_for_audit("Closing DB connection...")
+    db_connection.close()
+    # Archive file
+    s3_class = S3()
+    s3_class.copy_object(bucket, filename, event, start)
+    s3_class.delete_object(bucket, filename, event, start)
+    log_for_audit("Archived file {} to {}/archive/{}".format(filename, filename.split("/")[0], filename.split("/")[1]))
+    # Send Slack Notification
+    log_for_audit("Sending slack message...")
+    send_success_slack_message(event, start)
+    return "Cleanup Successful"
+
+
+def connect_to_database(env, event, start):
+    db = DB()
+    log_for_audit("Setting DB connection details")
+    if not db.db_set_connection_details(env, event, start):
+        log_for_error("Error DB Parameter(s) not found in secrets store.")
+        send_failure_slack_message(event, start)
+        raise ValueError("DB Parameter(s) not found in secrets store")
+    return db.db_connect(event, start)
+
+
+def retrieve_file_from_bucket(bucket, filename, event, start):
+    log_for_audit("Looking in {} for {} file".format(bucket, filename))
+    s3_bucket = S3()
+    return s3_bucket.get_object(bucket, filename, event, start)
